Little fix in xos solver.

+++ hermes_common/solver/nox.cpp
@@ -117,7 +117,7 @@ bool NoxProblemInterface::computeF(const Epetra_Vector &x, Epetra_Vector &f, Fil
   delete [] coeff_vec;
 
   for(unsigned int i = 0; i < rhs.length(); i++)
-    f.ReplaceGlobalValue(0, i, rhs.get(i));
+    f.ReplaceGlobalValue(0, (int)i, rhs.get(i));
   
   return true;
 }
